// src/user/models/user-role.ts
export enum UserRole {
  USER = 'USER',
  ADMIN = 'ADMIN',
  CUSTOMER = 'CUSTOMER',
  PROVIDER_COMPANY_ADMIN = 'PROVIDER_COMPANY_ADMIN',
  PROVIDER_SCHOOL_ADMIN = 'PROVIDER_SCHOOL_ADMIN',
  PROVIDER_INVESTOR_ADMIN = 'PROVIDER_INVESTOR_ADMIN',
  PARTNER = 'PARTNER',
  PROVIDER_COMPANY_MOD = 'PROVIDER_COMPANY_MOD',
  PROVIDER_SCHOOL_MOD = 'PROVIDER_SCHOOL_MOD',
  GUEST = 'GUEST',
}